requirements.txt
# DocMind AI — Requirements

streamlit>=1.32.0

# LangChain
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.2
langchain-classic>=1.0.2

# ── Local mode
langchain-ollama>=0.2.0

# ── Cloud mode (Groq)
langchain-groq>=0.1.0
groq>=0.9.0

# Vector DB + Embeddings
chromadb>=0.5.0
sentence-transformers>=2.7.0

# Document loading
pypdf>=4.1.0
beautifulsoup4>=4.12.3
requests>=2.31.0
html2text>=2024.2.26

# Export
reportlab>=4.1.0
python-docx>=1.1.0

# Env file support
python-dotenv>=1.0.0